\name{SNPMaP-internal}
\title{Internal SNPMaP objects}
\alias{.checkOpen}
\alias{.gws}
\alias{.isCel}
\alias{.makefile}
\alias{.meanMismatch}
\alias{.mm}
\alias{.ra}
\alias{extrA}
\alias{mate}
\alias{sca}
\alias{scaleq}
\alias{suco}
\description{Internal SNPMaP objects.}
\details{These are not to be called by the user.}
\keyword{internal}
